Increase whitelist of false detections in SdkCoreApiSurfaceTest
